Refactor ddpm_v3 structure and add tutorial
